Move out of utils/extensionUtils.ts
Because it can't be strict yet
