+++ src/modDefinitions/modDefinitions.test.tsx
@@ -29,8 +29,7 @@ import { appApiMock } from "@/testUtils/appApiMock";
 import { defaultModDefinitionFactory } from "@/testUtils/factories/modDefinitionFactories";
 import extensionsSlice from "@/store/extensionsSlice";
 
-jest.mock("@/background/navigation");
-
+jest.mock("@/utils/tabUtils");
 jest.mock("@/components/ConfirmationModal", () => ({
   ...jest.requireActual("@/components/ConfirmationModal"),
   useModals: () => ({
