+++ src/pageEditor/hooks/useSaveMod.test.ts
@@ -33,8 +33,7 @@ import extensionsSlice from "@/store/extensionsSlice";
 const modId = validateRegistryId("@test/mod");
 
 jest.mock("@/utils/notify");
-jest.mock("@/background/navigation");
-
+jest.mock("@/utils/tabUtils");
 jest.mock("@/components/ConfirmationModal", () => ({
   __esModule: true,
   useModals: () => ({
